Makes loading of xacml templates relative to source code.
The current loading is relative to the working dir of the
process. If the process is not started from the source
code dir, the templates are not loaded and the requests
to the keypass always fail. There are instalations
where the init.d script doesn't not execute the "cd"
to the source dir.

The templates are really part of the code and not a configurable
feature to the operators, so it is changed to always load
the templates relative to the sources regardless of the
working directory location.